# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lấy danh sách tất cả thành phẩm",
                "operationId": "7ca3593d05653d957ae7be1ac648ca44",
                "responses": {
                    "200": {
                        "description": "Danh sách thành phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Chai nhựa HDPE 1 lít xanh"
                                            },
                                            "category_name": {
                                                "type": "string",
                                                "example": "Chai nhựa PET"
                                            },
                                            "color_name": {
                                                "type": "string",
                                                "example": "Xanh"
                                            },
                                            "unit": {
                                                "type": "string",
                                                "example": "chai"
                                            },
                                            "quantity": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "product_img": {
                                                "type": "string",
                                                "example": null,
                                                "nullable": true
                                            },
                                            "status": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-21T15:34:21.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": null,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không có thành phẩm nào",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy danh sách thành phẩm"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Hiện tại chưa có thành phẩm nào"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi khi lấy danh sách thành phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy danh sách thành phẩm"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message here"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/product/add": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Thêm mới thành phẩm",
                "operationId": "4a27db21910050582b444b2a3d49831c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "unit",
                                    "quantity",
                                    "category_id",
                                    "color_id",
                                    "shelf_id",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nhựa HDPE"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "bao"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "color_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "material_img": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "image.jpg"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thêm thành phẩm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thêm thành phẩm thành công"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Thêm thành phẩm thất bại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thêm thành phẩm thất bại"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/product/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lấy thông tin thành phẩm theo ID",
                "operationId": "4576535dfeb323e297ec83ea67c18fce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của thành phẩm cần lấy thông tin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin thành phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Xô nhựa HDPE 1 lít đỏ"
                                        },
                                        "category_name": {
                                            "type": "string",
                                            "example": "Xô nhựa HDPE"
                                        },
                                        "color_name": {
                                            "type": "string",
                                            "example": "Đỏ"
                                        },
                                        "shelf_name": {
                                            "type": "string",
                                            "example": "Kệ 2"
                                        },
                                        "warehouse_name": {
                                            "type": "string",
                                            "example": "Kho thành phẩm 1"
                                        },
                                        "unit": {
                                            "type": "string",
                                            "example": "chai"
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "product_img": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-09-25T03:42:10.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy thành phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy thông tin thành phẩm"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Không tìm thấy thành phẩm"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi khi lấy thông tin thành phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy thông tin thành phẩm"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}